arch: move date helpers to shared
